requirements.txt
langchain
openai
huggingface_hub
python.dotenv
streamlit
langchain-openai
langchain-community
langchain-cli
langsmith
beautifulsoup4
faiss-cpu
langchainhub
langserve[all]
docarray
tiktoken
fastapi
uvicorn
chromadb
bs4
unstructured
astrapy
langchain-astradb